[FLINK-20006] Fix the unstable FileSinkITCase
